refactor: move canvas page widgets into canvas feature
